address lowercase bug and add tests
