+++ auth/test/test_auth_utils.py
@@ -1,6 +1,6 @@
 import pytest
 
-from auth.auth_utils import validate_credentials_secret_name_input
+from auth.auth_utils import is_valid_username, validate_credentials_secret_name_input
 from auth.exceptions import AuthUserError
 
 
@@ -62,3 +62,32 @@ def test_validate_credentials_secret_name_input_valid(input_secret_name):
 def test_validate_credentials_secret_name_input_invalid(input_secret_name):
     with pytest.raises(AuthUserError):
         validate_credentials_secret_name_input(input_secret_name)
+
+
+@pytest.mark.parametrize("input_username", ['abc', 'abc-def', 'a3a', 'a3a-3a', '123'])
+def test_is_valid_username_valid(input_username):
+    assert is_valid_username(input_username)
+
+
+@pytest.mark.parametrize(
+    "input_username",
+    [
+        'ABC',
+        'abc!',
+        'a3a.3a',
+        'abc_def',
+        'abc!def',
+        'abc.def.',
+        'abc.def',
+        'abc.def-ghi',
+        'abc-def.ghi',
+        'abc--def',
+        'abc.def--ghi',
+        'abc.def-.ghi',
+        '.abc',
+        'abc.',
+        'abc..def',
+    ],
+)
+def test_is_valid_username_invalid(input_username):
+    assert not is_valid_username(input_username)
